requirements.txt
google-api-python-client
yt-dlp
feedgen
python-dateutil
feedparser
